fix: move get_tenant_id_by_user_id to tenant repo
